update README and more...
