+++ serpapi/error.py
@@ -1,3 +1,3 @@
-"""error related classes"""
+"""wrap any serpapi related exception"""
 class SerpApiException(Exception):
     """custom exception for this module"""
