+++ tests/example_search_baidu_test.py
@@ -1,20 +1,21 @@
-# test Google Scholar search engine
+# Example: baidu search engine
 import unittest
 import os
 import serpapi
 
-class TestGoogleScholar(unittest.TestCase):
+class TestBaidu(unittest.TestCase):
 
   @unittest.skipIf((os.getenv("API_KEY") == None), "no api_key provided")
-  def test_search_google_scholar(self):
+  def test_search_baidu(self):
     client = serpapi.Client({
-        'engine': 'google_scholar',
+        'engine': 'baidu',
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'q': 'coffee', 
+        'q': 'coffee',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['organic_results'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
