+++ tests/example_search_ebay_test.py
@@ -1,20 +1,21 @@
-# test Bing search engine
+# Example: ebay search engine
 import unittest
 import os
 import serpapi
 
-class TestBing(unittest.TestCase):
+class TestEbay(unittest.TestCase):
 
   @unittest.skipIf((os.getenv("API_KEY") == None), "no api_key provided")
-  def test_search_bing(self):
+  def test_search_ebay(self):
     client = serpapi.Client({
-        'engine': 'bing',
+        'engine': 'ebay',
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'q': 'coffee', 
+        '_nkw': 'coffee',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['organic_results'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
