+++ tests/example_search_google_images_test.py
@@ -1,4 +1,4 @@
-# test Google Images search engine
+# Example: google_images search engine
 import unittest
 import os
 import serpapi
@@ -8,15 +8,16 @@ class TestGoogleImages(unittest.TestCase):
   @unittest.skipIf((os.getenv("API_KEY") == None), "no api_key provided")
   def test_search_google_images(self):
     client = serpapi.Client({
-        'engine': 'google',
+        'engine': 'google_images',
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'engine': 'google', 
-        'tbm': 'isch', 
-        'q': 'coffee', 
+        'engine': 'google_images',
+        'tbm': 'isch',
+        'q': 'coffee',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['images_results'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
