+++ tests/example_search_google_maps_test.py
@@ -1,4 +1,4 @@
-# test Google Maps search engine
+# Example: google_maps search engine
 import unittest
 import os
 import serpapi
@@ -12,11 +12,12 @@ class TestGoogleMaps(unittest.TestCase):
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'q': 'pizza', 
-        'll': '@40.7455096,-74.0083012,15.1z', 
-        'type': 'search', 
+        'q': 'pizza',
+        'll': '@40.7455096,-74.0083012,15.1z',
+        'type': 'search',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['local_results'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
