+++ tests/example_search_google_product_test.py
@@ -1,4 +1,4 @@
-# test Google Product search engine
+# Example: google_product search engine
 import unittest
 import os
 import serpapi
@@ -12,10 +12,11 @@ class TestGoogleProduct(unittest.TestCase):
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'q': 'coffee', 
-        'product_id': '4172129135583325756', 
+        'q': 'coffee',
+        'product_id': '4172129135583325756',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['product_results'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
