+++ tests/example_search_google_reverse_image_test.py
@@ -1,4 +1,4 @@
-# test Google Reverse Image search engine
+# Example: google_reverse_image search engine
 import unittest
 import os
 import serpapi
@@ -12,9 +12,10 @@ class TestGoogleReverseImage(unittest.TestCase):
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'image_url': 'https://i.imgur.com/5bGzZi7.jpg', 
+        'image_url': 'https://i.imgur.com/5bGzZi7.jpg',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['image_sizes'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
