+++ tests/example_search_google_test.py
@@ -1,21 +1,22 @@
-# test Google Search search engine
+# Example: google search engine
 import unittest
 import os
 import serpapi
 
-class TestGoogleSearch(unittest.TestCase):
+class TestGoogle(unittest.TestCase):
 
   @unittest.skipIf((os.getenv("API_KEY") == None), "no api_key provided")
-  def test_search_google_search(self):
+  def test_search_google(self):
     client = serpapi.Client({
         'engine': 'google',
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'q': 'coffee', 
-        'engine': 'google', 
+        'q': 'coffee',
+        'engine': 'google',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['organic_results'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
