+++ tests/example_search_yahoo_test.py
@@ -1,20 +1,21 @@
-# test Walmart search engine
+# Example: yahoo search engine
 import unittest
 import os
 import serpapi
 
-class TestWalmart(unittest.TestCase):
+class TestYahoo(unittest.TestCase):
 
   @unittest.skipIf((os.getenv("API_KEY") == None), "no api_key provided")
-  def test_search_walmart(self):
+  def test_search_yahoo(self):
     client = serpapi.Client({
-        'engine': 'walmart',
+        'engine': 'yahoo',
         'api_key': os.getenv("API_KEY")
       })
     data = client.search({
-        'query': 'coffee', 
+        'p': 'coffee',
     })
     self.assertIsNone(data.get('error'))
     self.assertIsNotNone(data['organic_results'])
-    # os.getenv("API_KEY") captures the secret user API available from http://serpapi.com
-  
+    # os.getenv("API_KEY") is your secret API Key
+    # copy/paste from [http://serpapi.com/dashboard] to your bash
+    # ```export API_KEY="your_secure_api_key"```
